dags/requirements.txt
apache-airflow==2.9.1
google-api-python-client
pandas
requests
beautifulsoup4
